requirements.txt
ruff
opencv-python
moviepy
simpleaudio
beautifulsoup4
pyyaml